# backend/src/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tayyibat Dietary System API — Dr. Dhia Al-Awadi
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ── FOODS ─────────────────────────────────────────────────────────────────
  /foods:
    get:
      operationId: listFoods
      tags: [foods]
      summary: List all foods with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [tayyib, khabeeth]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodListResponse"
...
  /foods/{id}:
    get:
      operationId: getFood
      tags: [foods]
      summary: Get a single food item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "404":
          description: Not found

  # ── TRACKER ───────────────────────────────────────────────────────────────
  /tracker/entries:
    get:
      operationId: listTrackerEntries
      tags: [tracker]
      summary: List tracker entries for a date range
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackerEntry"

    post:
      operationId: createTrackerEntry
      tags: [tracker]
      summary: Log a meal or habit entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackerEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerEntry"
  /tracker/entries/{id}:
    patch:
      operationId: updateTrackerEntry
      tags: [tracker]
      summary: Update a tracker entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackerEntryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackerEntry"

    delete:
      operationId: deleteTrackerEntry
      tags: [tracker]
      summary: Delete a tracker entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
...
  /tracker/weight:
    get:
      operationId: listWeightLogs
      tags: [tracker]
      summary: List weight log entries
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightLog"

    post:
      operationId: createWeightLog
      tags: [tracker]
      summary: Log a weight entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightLog"

  # ── PLANS ─────────────────────────────────────────────────────────────────
  /plans:
    get:
      operationId: listMealPlans
      tags: [plans]
      summary: List user's meal plans
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealPlan"

    post:
      operationId: createMealPlan
      tags: [plans]
      summary: Create a new meal plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealPlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlan"
  /plans/{id}:
    get:
      operationId: getMealPlan
      tags: [plans]
      summary: Get a specific meal plan with its entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlan"

    patch:
      operationId: updateMealPlan
      tags: [plans]
      summary: Update a meal plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealPlanUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlan"

    delete:
      operationId: deleteMealPlan
      tags: [plans]
      summary: Delete a meal plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # ── RECIPES ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── FOODS ─────────────────────────────────────────────────────────────
    Food:
      type: object
      required: [id, name, nameAr, status, category]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        status:
          type: string
          enum: [tayyib, khabeeth]
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    FoodListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Food"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    TrackerEntry:
      type: object
      required: [id, userId, type, date, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [meal, water, fasting, symptom]
        date:
          type: string
          format: date
        time:
          type: ["string", "null"]
        foodName:
          type: ["string", "null"]
        foodStatus:
          type: ["string", "null"]
        quantity:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        waterMl:
          type: ["integer", "null"]
        fastingHours:
          type: ["number", "null"]
        hungerLevel:
          type: ["integer", "null"]
        energyLevel:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    TrackerEntryInput:
      type: object
      required: [type, date]
      properties:
        type:
          type: string
          enum: [meal, water, fasting, symptom]
        date:
          type: string
          format: date
        time:
          type: string
        foodName:
          type: string
        foodStatus:
          type: string
          enum: [tayyib, khabeeth]
        quantity:
          type: string
        notes:
          type: string
        imageUrl:
          type: string
        waterMl:
          type: integer
        fastingHours:
          type: number
        hungerLevel:
          type: integer
          minimum: 1
          maximum: 10
        energyLevel:
          type: integer
          minimum: 1
          maximum: 10
    TrackerEntryUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [meal, water, fasting, symptom]
        date:
          type: string
          format: date
        time:
          type: string
        foodName:
          type: string
        foodStatus:
          type: string
        quantity:
          type: string
        notes:
          type: string
        imageUrl:
          type: string
        waterMl:
          type: integer
        fastingHours:
          type: number
        hungerLevel:
          type: integer
        energyLevel:
          type: integer
...
    WeightLog:
      type: object
      required: [id, userId, weightKg, date, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        weightKg:
          type: number
        date:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    WeightLogInput:
      type: object
      required: [weightKg, date]
      properties:
        weightKg:
          type: number
        date:
          type: string
          format: date
        notes:
          type: string

    # ── PLANS ─────────────────────────────────────────────────────────────
    MealPlan:
      type: object
      required: [id, userId, title, goal, startDate, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        titleAr:
          type: ["string", "null"]
        goal:
          type: string
          enum: [weight_loss, diabetes, energy, general]
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
        isActive:
          type: boolean
        notes:
          type: ["string", "null"]
        entries:
          type: array
          items:
            $ref: "#/components/schemas/MealPlanEntry"
        createdAt:
          type: string
          format: date-time
    MealPlanEntry:
      type: object
      required: [id, planId, dayOfWeek, mealTime, foodName]
      properties:
        id:
          type: integer
        planId:
          type: integer
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        mealTime:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        foodName:
          type: string
        quantity:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    MealPlanInput:
      type: object
      required: [title, goal, startDate]
      properties:
        title:
          type: string
        titleAr:
          type: string
        goal:
          type: string
          enum: [weight_loss, diabetes, energy, general]
        startDate:
          type: string
          format: date
        endDate:
          type: string
        notes:
          type: string
    MealPlanUpdate:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        goal:
          type: string
          enum: [weight_loss, diabetes, energy, general]
        startDate:
          type: string
          format: date
        endDate:
          type: string
        isActive:
          type: boolean
        notes:
          type: string

    # ── RECIPES ───────────────────────────────────────────────────────────
...